refactor: User 도메인 영역의 API를 Creator 영역으로 이동
